fix(dashboard): humanize overview attention reason/action labels via shared SSOT
The overview attention queue rendered the backend `attention_reason` and
`next_human_action` wire codes raw (e.g. `inspect_blocker_before_resume`,
`inspect_latest_error`), while the keeper detail alert strip already mapped
the same vocabulary to Korean labels. The prototype shows humanized actions
("원인 확인 후 재개", "최근 오류 확인").

Extract the closed-sum humanization (unions, label records, canonical folds,
unknown-token warn) into src/lib/keeper-attention-labels.ts as the single
source of truth. keeper-detail-alert-strip imports from it (behavior-identical;
its canonical-folding tests still pass) and overview maps through it too.
Unknown codes — including colon-composites like `completion_contract_result:*`
(backend reason leakage tracked separately) — fall back to the raw token and
warn in dev rather than being dropped.
